video/clipboard: the clipboard's session-epoch copy is u64 — a composition defect between APPCLIP (b208b5cd, u32 copy) and SECLOGIN M5 (9081402e, the epoch widened to u64 on both arches), found by gate11's strict check at 35b4f069
Neither executor saw the other's tree: APPCLIP's base predates SECLOGIN's widening and SECLOGIN's base predates APPCLIP. gate11 (the flight-12 image build) red every login-carrying cfg leg on both arches with one error, clipboard.rs:121 "expected u32, found u64". The fix widens the copy (the struct field and the shim's return type); it never narrows the epoch — SECLOGIN's u64 exists because a u32 wraps. Nothing else in the file is typed on the width: the fixture prints `epoch={}` and compares `c.epoch != live`. gate11 is relaunched on this commit; ledger-check rc=0 before and after.

+++ unaos/crates/kernel/src/video/clipboard.rs
@@ -99,7 +99,7 @@ struct Clip {
     buf: [u8; CLIP_CAP],
     len: usize,
     /// The session epoch this content was [`set`] under.
-    epoch: u32,
+    epoch: u64, // u64 since SECLOGIN M5 widened the session epoch (fold 9081402e); the seat widened this copy at the fold seam 2026-09-23 — the epoch is never narrowed
     /// Has anything ever been set? `false` means "empty by construction", not "cleared".
     stamped: bool,
 }
@@ -115,7 +115,7 @@ static CLIP: Mutex<Clip> = Mutex::new(Clip {
 /// function is private to a file this arc does not edit, so the shape is duplicated and not the
 /// value). `0` where there is no `login` feature or no EL0 regime to carry a session — see the
 /// module header on why that does not weaken the ownership check.
-fn session_epoch() -> u32 {
+fn session_epoch() -> u64 {
     #[cfg(all(feature = "login", any(target_arch = "x86_64", feature = "aarch64_el0")))]
     {
         return crate::arch::syscall::session_epoch();
